mudanças

+++ src/main/java/com/aceleramaker/projeto/blogpessoal/service/PostagemServiceImpl.java
@@ -5,14 +5,20 @@ import com.aceleramaker.projeto.blogpessoal.controller.schema.CriarPostagemDTO;
 import com.aceleramaker.projeto.blogpessoal.controller.schema.FiltrosPostagemDTO;
 import com.aceleramaker.projeto.blogpessoal.controller.schema.PostagemDTO;
 import com.aceleramaker.projeto.blogpessoal.model.Postagem;
+import com.aceleramaker.projeto.blogpessoal.model.UsuarioLogin;
 import com.aceleramaker.projeto.blogpessoal.model.exception.EntidadeNaoEncontradaException;
+import com.aceleramaker.projeto.blogpessoal.model.exception.OperacaoNaoAutorizadaException;
 import com.aceleramaker.projeto.blogpessoal.repository.PostagemRepository;
 import com.aceleramaker.projeto.blogpessoal.repository.TemaRepository;
 import com.aceleramaker.projeto.blogpessoal.repository.UsuarioRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class PostagemServiceImpl implements PostagemService {
@@ -31,7 +37,11 @@ public class PostagemServiceImpl implements PostagemService {
     }
 
     public PostagemDTO criar(CriarPostagemDTO dto) {
-        var usuario = usuarioRepository.findById(dto.usuarioId())
+
+        var usuarioLogado = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        var usuarioLogin = (UsuarioLogin) usuarioLogado;
+
+        var usuario = usuarioRepository.findById(usuarioLogin.getUsuario().getId())
                 .orElseThrow(() -> new EntidadeNaoEncontradaException("Usuario"));
 
         var tema = temaRepository.findById(dto.temaId())
@@ -47,9 +57,16 @@ public class PostagemServiceImpl implements PostagemService {
         var postagem = postagemRepository.findById(postagemId)
                 .orElseThrow(() -> new EntidadeNaoEncontradaException("Postagem"));
 
-        if (dto.titulo() != null) postagem.setTitulo(postagem.getTitulo());
+        var usuarioLogado = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        var usuarioLogin = (UsuarioLogin) usuarioLogado;
+        
+        if (!postagem.getUsuario().getId().equals(usuarioLogin.getUsuario().getId())) {
+            throw new OperacaoNaoAutorizadaException("Não é possível atualizar postagens de outros usuários");
+        }
 
-        if (dto.texto() != null) postagem.setTexto(postagem.getTexto());
+        if (dto.titulo() != null) postagem.setTitulo(dto.titulo());
+
+        if (dto.texto() != null) postagem.setTexto(dto.texto());
 
         if (dto.temaId() != null) {
             var tema = temaRepository.findById(dto.temaId())
@@ -64,7 +81,37 @@ public class PostagemServiceImpl implements PostagemService {
         return postagemRepository.buscarComFiltros(filtros, pageable).map(PostagemDTO::new);
     }
 
+    public Page<PostagemDTO> buscarPostagensPorUsuarioLogado(Pageable pageable) {
+        var usuarioLogado = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        var usuarioLogin = (UsuarioLogin) usuarioLogado;
+
+        return postagemRepository.buscarPostagensPorUsuario(usuarioLogin.getUsuario().getId(), pageable)
+                .map(PostagemDTO::new);
+    }
+
+    public List<PostagemDTO> listarUltimosPosts() {
+        return postagemRepository.listarUltimosPosts().stream()
+                .map(PostagemDTO::new)
+                .collect(Collectors.toList());
+    }
+
+    public PostagemDTO buscarPorId(Long id) {
+        return postagemRepository.findById(id)
+                .map(PostagemDTO::new)
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Postagem"));
+    }
+
     public void deletar(Long id) {
+        var postagem = postagemRepository.findById(id)
+                .orElseThrow(() -> new EntidadeNaoEncontradaException("Postagem"));
+
+        var usuarioLogado = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        var usuarioLogin = (UsuarioLogin) usuarioLogado;
+        
+        if (!postagem.getUsuario().getId().equals(usuarioLogin.getUsuario().getId())) {
+            throw new OperacaoNaoAutorizadaException("Não é possível deletar postagens de outros usuários");
+        }
+        
         postagemRepository.deleteById(id);
     }
 }
